ci: run exhaustive tenant migration tests on demand
