+++ internal/devtool/catalog.go
@@ -45,6 +45,7 @@ var gates = []Gate{
 	{ID: "go-staticcheck", Description: "Run pinned Staticcheck", CIGroup: "go-checks", Command: []string{"go", "run", "honnef.co/go/tools/cmd/staticcheck@v" + ToolVersion("staticcheck"), "./..."}, Profiles: []string{"pr", "full"}, Paths: []string{"*.go", "go.mod", "go.sum"}, Weight: 2, Timeout: "15m"},
 	{ID: "go-race-heavy", Description: "Run race tests for the canonical heavy Go package shard", CIGroup: "go-race", Command: []string{"./hk", "ci", "race", "--shard", "heavy"}, Profiles: []string{"pr", "full"}, Paths: []string{"*.go", "*.sql", "go.mod", "go.sum"}, Weight: 4, Timeout: "30m"},
 	{ID: "go-race-rest", Description: "Run race tests for the complementary Go package shard", CIGroup: "go-race", Command: []string{"./hk", "ci", "race", "--shard", "rest"}, Profiles: []string{"pr", "full"}, Paths: []string{"*.go", "*.sql", "go.mod", "go.sum"}, Weight: 4, Timeout: "30m"},
+	{ID: "default-tenant-migration-acceptance", Description: "Run exhaustive default-tenant split fault and large-fingerprint acceptance tests", Command: []string{"env", "HITKEEP_DEFAULT_TENANT_MIGRATION_ACCEPTANCE=1", "go", "test", "./internal/database", "-run", "Test(DefaultTenantSplitFaultBoundariesResume|SplitFingerprintSurvivesLargeParquetRoundTrip)$"}, Profiles: []string{"full"}, Paths: []string{"internal/database/", "internal/devtool/", ".github/workflows/"}, Weight: 4, Timeout: "30m"},
 	{ID: "mcp-audit", Description: "Audit the production MCP protocol surface", CIGroup: "production-mcp", Command: []string{"go", "test", "./internal/mcpserver", "-run", "TestMCP.*Audit"}, Profiles: []string{"pr", "full"}, Paths: []string{"internal/mcpserver/", "internal/analyticstools/", "skills/", "server.json"}, Weight: 2, Timeout: "10m"},
 	{ID: "mcp-schema", Description: "Validate MCP registry metadata", CIGroup: "production-mcp", Command: []string{"tests/scripts/mcp-audit.sh", "--schema-only"}, Profiles: []string{"pr", "full"}, Paths: []string{"internal/mcpserver/", "tests/scripts/mcp-audit.sh", "server.json"}, Weight: 1, Timeout: "5m"},
 	{ID: "developer-mcp", Description: "Validate the developer CLI/MCP contract and production binary boundary", CIGroup: "go-checks", Command: []string{"go", "test", "./cmd/hk", "./internal/devtool/devmcp"}, Profiles: []string{"pr", "full"}, Paths: []string{"cmd/hk/", "cmd/hitkeep/", "internal/devtool/", "hk", "go.mod", "go.sum"}, Weight: 1, Timeout: "5m"},
