+++ internal/server/admin/system_handlers_test.go
@@ -57,16 +57,7 @@ func setupSystemTestEnv(t *testing.T) (*handler, *database.Store, *database.Tena
 	if err := store.UpdateInstanceRole(context.Background(), adminUserID, auth.InstanceAdmin, ownerUserID); err != nil {
 		t.Fatalf("promote admin: %v", err)
 	}
-	if err := store.Close(); err != nil {
-		t.Fatalf("close control before split: %v", err)
-	}
-	if err := database.RunDefaultTenantSplit(context.Background(), sharedPath, basePath); err != nil {
-		t.Fatalf("split default tenant: %v", err)
-	}
-	store = database.NewStore(sharedPath)
-	if err := store.Connect(); err != nil {
-		t.Fatalf("reopen control after split: %v", err)
-	}
+	prepareEmptySystemTestDataPlane(t, store, basePath)
 
 	tenantStores := database.NewTenantStoreManager(store, basePath, database.WithTenantDataPlane(true))
 	t.Cleanup(func() { _ = tenantStores.Close() })
@@ -96,6 +87,43 @@ func setupSystemTestEnv(t *testing.T) (*handler, *database.Store, *database.Tena
 	}
 
 	return &handler{ctx: ctx}, store, tenantStores, ownerUserID, adminUserID, regularUserID
+}
+
+// prepareEmptySystemTestDataPlane builds the already-split topology needed by
+// system handler tests. Migration crash/rewrite behavior is covered in the
+// database package; repeating the physical control rewrite for every handler
+// test made the race shard spend minutes rebuilding the same empty catalogs.
+func prepareEmptySystemTestDataPlane(t *testing.T, control *database.Store, dataPath string) {
+	t.Helper()
+	ctx := context.Background()
+	defaultTenantID, err := control.GetDefaultTenantID(ctx)
+	if err != nil {
+		t.Fatalf("resolve default tenant: %v", err)
+	}
+	tenantDir := filepath.Join(dataPath, "tenants", defaultTenantID.String())
+	if err := os.MkdirAll(tenantDir, 0o700); err != nil {
+		t.Fatalf("create default tenant directory: %v", err)
+	}
+	tenant := database.NewStore(filepath.Join(tenantDir, "hitkeep.db"))
+	if err := tenant.Connect(); err != nil {
+		t.Fatalf("connect default tenant test store: %v", err)
+	}
+	if err := tenant.MigrateTenant(ctx); err != nil {
+		_ = tenant.Close()
+		t.Fatalf("migrate default tenant test store: %v", err)
+	}
+	if err := tenant.Close(); err != nil {
+		t.Fatalf("close default tenant test store: %v", err)
+	}
+	if _, err := control.DB().ExecContext(ctx, `
+		INSERT INTO data_migrations (name, applied_at)
+		VALUES
+			('default_tenant_split_v1', now()),
+			('default_tenant_split_compacted_v1', now())
+		ON CONFLICT (name) DO NOTHING
+	`); err != nil {
+		t.Fatalf("mark empty test data plane as split: %v", err)
+	}
 }
 
 func TestHandleGetSystem(t *testing.T) {
